Refactoring kill-yank module structure
