+++ src/kill-yank/kill-ring-entity/clipboard-text.ts
