+++ src/kill-yank/kill-ring-entity/kill-ring-entity.ts
