+++ src/test/commands/kill-yank/kill-ring-yank.test.ts
@@ -4,7 +4,7 @@ import * as vscode from "vscode";
 import {Position, Range, Selection} from "vscode";
 import { moveCommandIds } from "../../../commands/move";
 import { EmacsEmulator } from "../../../emulator";
-import { KillRing } from "../../../kill-ring";
+import { KillRing } from "../../../kill-yank/kill-ring";
 import { assertTextEqual, cleanUpWorkspace, clearTextEditor, setupWorkspace} from "../../utils";
 
 suite("kill, yank, yank-pop", () => {
