+++ src/test/commands/kill-yank/kill-word.test.ts
@@ -1,6 +1,6 @@
 import { TextEditor } from "vscode";
 import { EmacsEmulator } from "../../../emulator";
-import { KillRing } from "../../../kill-ring";
+import { KillRing } from "../../../kill-yank/kill-ring";
 import {
     assertCursorsEqual, assertTextEqual,
     cleanUpWorkspace, clearTextEditor,
